+++ tests/Serendipity/Metatron/Command/Diag/ConfigCommandTest.php
@@ -1,15 +1,15 @@
 <?php
 
-require_once(S9Y_INCLUDE_PATH . 'metatron/tests/AbstractTest.php');
+namespace Serendipity\Metatron\Command\Diag;
 
 use Serendipity\Metatron\Application;
-use Serendipity\Metatron\Command\Diag\ConfigCommand;
+use Serendipity\Metatron\Command\PHPUnit\TestCase;
 use Symfony\Component\Console\Tester\CommandTester;
 
 /**
  * Class ConfigCommandTest
  */
-class ConfigCommandTest extends AbstractTest
+class ConfigCommandTest extends TestCase
 {
     /**
      * @covers ConfigCommand::execute